+++ netket/experimental/operator/__init__.py
@@ -17,8 +17,8 @@ __all__ = ["FermionOperator2nd", "FermionOperator2ndJax"]
 
 from .pyscf import from_pyscf_molecule
 from ._particle_number_conserving_fermionic import (
-    ParticleNumberConservingFermioperator2ndJax,
-    ParticleNumberConservingFermioperator2ndSpinJax,
+    ParticleNumberConservingFermioperator2nd,
+    ParticleNumberAndSpinConservingFermioperator2nd,
     FermiHubbardJax,
 )