+++ netket/experimental/operator/pyscf.py
@@ -21,8 +21,8 @@ from netket.hilbert import SpinOrbitalFermions
 from netket.utils.optional_deps import import_optional_dependency
 from netket.operator import FermionOperator2nd, FermionOperator2ndJax
 from ._particle_number_conserving_fermionic import (
-    ParticleNumberConservingFermioperator2ndJax,
-    ParticleNumberConservingFermioperator2ndSpinJax,
+    ParticleNumberConservingFermioperator2nd,
+    ParticleNumberAndSpinConservingFermioperator2nd,
 )
 
 from ._pyscf_utils import arrays_to_terms, TV_from_pyscf_molecule
@@ -66,7 +66,7 @@ def from_pyscf_molecule(
     mo_coeff: np.ndarray | None = None,
     *,
     cutoff: float = 1e-11,
-    implementation: DiscreteOperator = ParticleNumberConservingFermioperator2ndSpinJax,
+    implementation: DiscreteOperator = ParticleNumberAndSpinConservingFermioperator2nd,
 ) -> DiscreteOperator:
     r"""
     Construct a netket operator encoding the electronic hamiltonian of a pyscf
@@ -122,7 +122,7 @@ def from_pyscf_molecule(
         implementation: The particular implementation to use for the operator.
             Different fermionic operator implementation might have different
             performances. Defaults to
-            :class:`netket.experimental.operator.ParticleNumberConservingFermioperator2ndSpinJax` (this might
+            :class:`netket.experimental.operator.ParticleNumberAndSpinConservingFermioperator2nd` (this might
             change in the future).
 
     Returns:
@@ -149,8 +149,8 @@ def from_pyscf_molecule(
         # TODO maybe run setup and set _max_conn_size here estimating it analytially
         return ha
     elif implementation in [
-        ParticleNumberConservingFermioperator2ndJax,
-        ParticleNumberConservingFermioperator2ndSpinJax,
+        ParticleNumberConservingFermioperator2nd,
+        ParticleNumberAndSpinConservingFermioperator2nd,
     ]:
         return implementation.from_pyscf_molecule(molecule, mo_coeff)
     else: